requirements.txt
python-dotenv~=1.0.1
openai~=1.35.7
streamlit~=1.36.0
mediawikiapi~=1.2
nltk~=3.8.1
tiktoken~=0.9.0